+++ plugins/events-correlation-engine/src/main/java/org/opensearch/plugin/correlation/core/index/codec/CorrelationCodecVersion.java
@@ -11,8 +11,8 @@ package org.opensearch.plugin.correlation.core.index.codec;
 import org.apache.lucene.codecs.Codec;
 import org.apache.lucene.codecs.lucene99.Lucene99Codec;
 import org.opensearch.index.mapper.MapperService;
-import org.opensearch.plugin.correlation.core.index.codec.correlation950.CorrelationCodec;
-import org.opensearch.plugin.correlation.core.index.codec.correlation950.PerFieldCorrelationVectorsFormat;
+import org.opensearch.plugin.correlation.core.index.codec.correlation990.CorrelationCodec;
+import org.opensearch.plugin.correlation.core.index.codec.correlation990.PerFieldCorrelationVectorsFormat;
 
 import java.util.Optional;
 import java.util.function.BiFunction;
