+++ plugins/events-correlation-engine/src/main/resources/META-INF/services/org.apache.lucene.codecs.Codec
@@ -1,1 +1,1 @@
-org.opensearch.plugin.correlation.core.index.codec.correlation950.CorrelationCodec
+org.opensearch.plugin.correlation.core.index.codec.correlation990.CorrelationCodec
